refactor-restructured-bl: decoupled responsibilites from controller to service and api levels.
